edit log items
